changing the transfer request columns names

+++ app/Http/Controllers/Dashboard/UserController.php
@@ -5,17 +5,15 @@ namespace App\Http\Controllers\Dashboard;
 use App\Http\Requests\Dashboard\User\StoreUserRequest;
 use App\Http\Requests\Dashboard\User\UpdateUserRequest;
 use App\Http\Resources\Dashboard\UserResource;
-
 use App\Models\User;
 use App\Traits\Responses;
-
 use Illuminate\Http\Response;
 use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
-    use Responses ;
+    use Responses;
 
     public function index()
     {
@@ -35,7 +33,7 @@ class UserController extends Controller
         $data['password'] = Hash::make('password');
         $data['security_level'] = $data['security_level'] ?? 4;
 
-        $user = User::create($data);
+        $user = User::on('sqlsrv_rms')->create($data);
 
         return $this->success(
             status: Response::HTTP_CREATED,
@@ -47,8 +45,17 @@ class UserController extends Controller
     /**
      * Display the specified user
      */
-    public function show(User $user)
+    public function show($id)
     {
+        $user = User::on('sqlsrv_rms')->find($id);
+
+        if (!$user) {
+            return $this->error(
+                status: Response::HTTP_NOT_FOUND,
+                message: 'User not found'
+            );
+        }
+
         return $this->success(
             status: Response::HTTP_OK,
             message: 'User retrieved successfully',
@@ -59,9 +66,18 @@ class UserController extends Controller
     /**
      * Update the specified user
      */
-    public function update(User $user , UpdateUserRequest $request)
+    public function update($id, UpdateUserRequest $request)
     {
-        $updateData = $request->only(['name', 'email', 'store_id', 'user_number', 'role' ]);
+        $user = User::on('sqlsrv_rms')->find($id);
+
+        if (!$user) {
+            return $this->error(
+                status: Response::HTTP_NOT_FOUND,
+                message: 'User not found'
+            );
+        }
+
+        $updateData = $request->only(['name', 'email', 'store_id', 'user_number', 'role']);
 
         $updateData = array_filter($updateData, fn($value) => !is_null($value));
 
@@ -79,8 +95,16 @@ class UserController extends Controller
     /**
      * Remove the specified user
      */
-    public function destroy(User $user)
+    public function destroy($id)
     {
+        $user = User::on('sqlsrv_rms')->find($id);
+        
+        if (!$user) {
+            return $this->error(
+                status: Response::HTTP_NOT_FOUND,
+                message: 'User not found'
+            );
+        }
         $user->delete();
 
         return $this->success(
